Split up some of the responsibilities in editlog.js into smaller domains; added autoloader and src folder.
